[SERVICE-565] Extract getContiguousEntities for Unit Testing
